Add helper to download the latest clangd snapshot
Note that no version check is performed, since it's impossble to make a
comparison using only commit hashes. So this patch keeps it simple and
only allows to download the latest version unconditionally.

The asset chooser function has been updated to avoid downloading the
indexing tools that have been present in the latest prereleases